Fix import paths and type exports after restructure
- Fix backend schema imports to use correct relative paths
- Export JobBatch interface from jobService for proper typing
- Update MCP API exports to include AI service functions
- Both backend and MCP now build successfully
