+++ backend/src/lib/ingestion/webScrape.ts
@@ -1,6 +1,6 @@
 import slug from 'slug';
 import { db } from '../db';
-import { libraries } from '../db/schema';
+import { libraries } from '../schema.js';
 import { eq } from 'drizzle-orm';
 import { WebScrapeSource } from '../types';
 import { sendEvent, closeConnection } from '../events';
