+++ backend/src/lib/jobs/service.ts
@@ -1,5 +1,5 @@
 import { jobService } from './jobService';
-import { EmbeddingJobPayload } from './jobService';
+import { EmbeddingJobPayload, JobBatch } from './jobService';
 
 export async function startCrawlJob(
   libraryName: string,
