+++ mcp/src/lib/api.ts
@@ -1,1 +1,2 @@
-export * from './libraries/search';
+export * from './libraries/search.js';
+export * from './ai/service.js';
